+++ trunk/RTCP/Cobalt/GPUProc/src/cuda/Kernels/UHEP_InvFIR_Kernel.cc
@@ -38,7 +38,7 @@ namespace LOFAR
       setArg(1, devFFTedData);
       setArg(2, devInvFIRfilterWeights);
 
-      size_t maxNrThreads, nrThreads;
+      unsigned maxNrThreads, nrThreads;
       //getWorkGroupInfo(queue.getInfo<CL_QUEUE_DEVICE>(), CL_KERNEL_WORK_GROUP_SIZE, &maxNrThreads);
       maxNrThreads = getAttribute(CU_FUNC_ATTRIBUTE_MAX_THREADS_PER_BLOCK);
       // round down to nearest power of two
